@@ -48,10 +48,10 @@ namespace AdventOfCode2022.Day2
         {
         }
 
-        protected override int RunPartOne() => GetAnswerPart1();
+        public override int RunPartOne() => GetAnswerPart1();
 
 
-        protected override int RunPartTwo() => GetAnswerPart2();
+        public override int RunPartTwo() => GetAnswerPart2();
 
         private int GetAnswerPart1()
         {
