@@ -5,39 +5,9 @@ namespace AdventOfCode2023;
 public record ScratchCard(int cardNumber, IEnumerable<int> WinningNumbers)
 { }
 
-public class ProgramDay4 : AdventOfCodeProgram<long>
+public class ProgramDay4(string? text = null) : AdventOfCodeProgram<long>(text)
 {
-    public ProgramDay4(string? text = null) : base(text)
-    {
-    }
-
-    [Theory]
-    [InlineData(@"Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
-Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
-Card 3:  1 21 53 59 44 | 69 82 63 72 16 21 14  1
-Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
-Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
-Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11",
-13)]
-    public override void RunTestsPartOne(string input, long expectedResult)
-    {
-        new ProgramDay4(input).RunPartOne().Should().Be(expectedResult);
-    }
-
-    [Theory]
-    [InlineData(@"Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
-Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
-Card 3:  1 21 53 59 44 | 69 82 63 72 16 21 14  1
-Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
-Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
-Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11",
-30)]
-    public override void RunTestsPartTwo(string input, long expectedResult)
-    {
-        new ProgramDay4(input).RunPartTwo().Should().Be(expectedResult);
-    }
-
-    protected override long RunPartOne()
+    public override long RunPartOne()
     {
         var total = Lines
         .Select(CreateScratchCards)
@@ -48,7 +18,7 @@ Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11",
         return (long)total;
     }
 
-    protected override long RunPartTwo()
+    public override long RunPartTwo()
     {
         var allScratchCards = Lines.Select(CreateScratchCards);
         var cardDictionary = allScratchCards.ToDictionary(c => c.cardNumber, c => 1);
